moved around the loader files into a loader folder with all things related to loader in it.
