+++ src/index.js
@@ -8,7 +8,7 @@ var EventEmitter = require("events");
 var Stream = require("stream");
 var es = require("event-stream");
 var deprecatedOptions = require("./deprecatedOptions")("bit-bundler");
-var loaderFactory = require("./loaderFactory");
+var loaderFactory = require("./loader/factory");
 var Bundler = require("./bundler");
 var Bundle = require("./bundle");
 var Context = require("./context");
