+++ src/loader/index.js
@@ -5,8 +5,8 @@ var utils = require("belty");
 var pullingDeps = require("pulling-deps");
 var resolvePath = require("bit-bundler-utils/resolvePath");
 var readFile = require("bit-bundler-utils/readFile");
-var pluginLoader = require("./pluginLoader");
-var logger = require("./logger").create("bundler/loader");
+var pluginLoader = require("../pluginLoader");
+var logger = require("../logger").create("bundler/loader");
 
 var moduleNotFoundError = buildError.bind(null, "Unable to find module");
 var moduleNotLoadedError = buildError.bind(null, "Unable to load module");
